implemented mentor feedback and changed position of class MarkerHighlightX and cleared comments
