+++ invesalius/data/visualization/marker_visualizer.py
@@ -2,9 +2,76 @@ import vtk
 
 import invesalius.constants as const
 import invesalius.data.coordinates as dco
+import invesalius.data.slice_ as sl
 import invesalius.session as ses
 from invesalius.data.markers.marker import MarkerType
+from invesalius.gui.widgets.canvas_renderer import CanvasHandlerBase
 from invesalius.pubsub import pub as Publisher
+
+
+class MarkerHighlightX(CanvasHandlerBase):
+    """A red 'X' mark drawn at a 3-D world position on the slice canvas.
+
+    The X mark is only visible on the slice where the marker was created,
+    ensuring precise spatial awareness of the marker's location.
+    """
+
+    _COLOUR = (255, 0, 0, 255)  # bright red for maximum visibility
+    _SIZE = 7  # half-width of X mark (14px total span)
+    _LINE_WIDTH = 2  # line thickness
+
+    def __init__(self, position=(0, 0, 0)):
+        super().__init__(parent=None)
+        self.position = position
+        self.layer = 99  # draw on top of everything
+        self.visible = True
+        self.sagittal_slice = None  # slice index for sagittal view
+        self.coronal_slice = None  # slice index for coronal view
+        self.axial_slice = None  # slice index for axial view
+
+    def draw_to_canvas(self, gc, canvas):
+        """Draw the X mark only on the slice where the marker was created."""
+        if not self.visible:
+            return
+
+        viewer = canvas.viewer
+
+        # Check if we're on the correct slice for this orientation
+        if viewer.slice_data is not None:
+            current_slice = viewer.slice_data.number
+
+            # Only draw if we're on the exact slice where the marker was created
+            if viewer.orientation == "AXIAL":
+                if self.axial_slice is None or current_slice != self.axial_slice:
+                    return
+            elif viewer.orientation == "CORONAL":
+                if self.coronal_slice is None or current_slice != self.coronal_slice:
+                    return
+            elif viewer.orientation == "SAGITAL":
+                if self.sagittal_slice is None or current_slice != self.sagittal_slice:
+                    return
+
+        # Project 3D marker position to 2D screen coordinates
+        px, py = self._3d_to_2d(canvas.evt_renderer, self.position)
+        scale = viewer.GetContentScaleFactor()
+        size = self._SIZE * scale
+
+        # Draw an X mark using canvas.draw_line() method
+        # Line from top-left to bottom-right
+        canvas.draw_line(
+            (px - size, py - size),
+            (px + size, py + size),
+            colour=self._COLOUR,
+            width=self._LINE_WIDTH,
+        )
+
+        # Line from top-right to bottom-left
+        canvas.draw_line(
+            (px + size, py - size),
+            (px - size, py + size),
+            colour=self._COLOUR,
+            width=self._LINE_WIDTH,
+        )
 
 
 class MarkerVisualizer:
